backend/requirements.txt
streamlit
requests
PyPDF2
newspaper3k
lxml
lxml_html_clean
